ApiWebSocketConnection does not use seperate thread for websocket
