+++ Api/QuantConnect.Api.csproj
@@ -107,7 +107,7 @@
     <Compile Include="Api.cs" />
     <Compile Include="ApiConnection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WebSocketConnection.cs" />
+    <Compile Include="ApiWebSocketConnection.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Common\QuantConnect.csproj">
