+++ Common/Data/BaseData.cs
@@ -14,6 +14,10 @@
 */
 
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using Newtonsoft.Json;
+using QuantConnect.Logging;
 using QuantConnect.Util;
 
 namespace QuantConnect.Data
@@ -301,5 +305,41 @@ namespace QuantConnect.Data
         {
             throw new InvalidOperationException("Please implement GetSource(SubscriptionDataConfig, DateTime, bool) on your custom data type: " + GetType().Name);
         }
+
+        /// <summary>
+        /// Deserialize the message from the data server
+        /// </summary>
+        /// <param name="serialized">The data server's message</param>
+        /// <returns>An enumerable of base data, if unsuccessful, returns an empty enumerable</returns>
+        public static IEnumerable<BaseData> DeserializeMessage(string serialized)
+        {
+            try
+            {
+                var deserialized = JsonConvert.DeserializeObject(serialized, JsonSerializerSettings);
+
+                var enumerable = deserialized as IEnumerable<BaseData>;
+                if (enumerable != null)
+                {
+                    return enumerable;
+                }
+
+                var data = deserialized as BaseData;
+                if (data != null)
+                {
+                    return new[] { data };
+                }
+            }
+            catch (Exception err)
+            {
+                Log.Error("BaseData.DeserializeMessage(): {0}", err);
+            }
+
+            return Enumerable.Empty<BaseData>();
+        }
+
+        private static readonly JsonSerializerSettings JsonSerializerSettings = new JsonSerializerSettings
+        {
+            TypeNameHandling = TypeNameHandling.All
+        };
     }
 }
